motiKocApp(tercih-asistani)/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
audio-recorder-streamlit==0.0.10
blinker==1.8.2
blis==1.0.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
comtypes==1.4.8
confection==0.1.5
cymem==2.0.8
dataclasses==0.6
distro==1.9.0
entrypoints==0.4
filelock==3.16.1
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2024.10.0
fuzzywuzzy==0.18.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.26.1
idna==3.10
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
json-logging==1.3.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langcodes==3.4.1
language_data==1.2.0
Levenshtein==0.26.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.10
narwhals==1.10.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.8.1
numpy==2.0.2
openai==0.28.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
plotly==5.24.1
preshed==3.0.9
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.0
pyttsx3==2.98
pytz==2024.2
pywin32==308
PyYAML==6.0.2
qrcode==8.0
rank-bm25==0.2.2
RapidFuzz==3.10.0
referencing==0.35.1
regex==2024.9.11
reportlab==4.2.5
requests==2.32.3
rich==13.9.3
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
setuptools==75.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
streamlit==1.39.0
streamlit-calendar==1.2.0
streamlit-option-menu==0.4.0
sympy==1.13.1
tenacity==9.0.0
thinc==8.3.2
tokenizers==0.20.1
toml==0.10.2
toolz==1.0.0
torch==2.5.0
tornado==6.4.1
tqdm==4.66.5
transformers==4.45.2
trtokenizer==0.0.3
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wasabi==1.1.3
watchdog==5.0.3
weasel==0.4.1
wheel==0.44.0
wrapt==1.16.0
yarl==1.16.0
zeyrek==0.1.3
